fzf: Spicing up staging files
I modified the function (from PatrickF1/fzf.fish[^1]) that uses fzf to
view staged files to escape parenthesis (which is necessary for working
in NextJS) and adds "git add" before the list of files to quickly stage
them. I bound the new function to the same key bind as the old one
(Ctrl+s).