[INSTALL] fix for homebrew, drop fisher
When Homebrew is installed the first time, it needs to be added to
$PATH. The ternary also didn't work the way I thought it would, so I'm
separating it.

The problem with the fisher install is that this is a Bash script, and
so the `source` syntax doesn't work. Also the install script needs to be
run in fish so it may not make sense when running this install script.